+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased changes
 
 * migrate to GRPC queries ([#157](https://github.com/provenance-io/provwasm/issues/157))
+* add `MetadataAddress` type for encoding/decoding metadata
+  addresses ([#161](https://github.com/provenance-io/provwasm/pull/161))
 
 ## Releases